requirements.txt
addict==2.4.0
# -e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
ftfy==6.3.0
importlib_metadata==8.5.0
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.1.105
# nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
packaging==24.1
pillow==10.2.0
platformdirs==4.3.6
-e git+https://github.com/cocodataset/cocoapi@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#egg=pycocotools&subdirectory=PythonAPI
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
six==1.16.0
sympy==1.13.1
termcolor==2.5.0
terminaltables==3.1.10
tomli==2.0.2
# torch==2.2.1+cu121
# torchaudio==2.2.1+cu121
# torchvision==0.17.1+cu121
tqdm==4.66.5
triton==2.2.0
typing_extensions==4.9.0
wcwidth==0.2.13
yapf==0.40.2
zipp==3.20.2
mmcv==1.6.0
psutil==7.0.0
mmengine==0.10.7
# torch                    2.2.1+cu121
# torchaudio               2.2.1+cu121
# torchvision              0.17.1+cu121
# pip install torch==2.2.1 torchvision==0.17.1 torchaudio==2.2.1 --index-url https://download.pytorch.org/whl/cu121